% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.JULES.R
\name{detect.timestep}
\alias{detect.timestep}
\title{Detect timestep of JULES met files}
\usage{
detect.timestep(met.dir, met.regexp)
}
\arguments{
\item{met.dir}{}

\item{met.regexp}{}
}
\description{
Detect timestep of JULES met files
}
