% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.JULES.R
\name{model2netcdf.JULES}
\alias{model2netcdf.JULES}
\title{Code to convert JULES output into netCDF format}
\usage{
model2netcdf.JULES(outdir)
}
\arguments{
\item{outdir}{Location of model output}
}
\description{
Convert MODEL output into the PEcAn standard
}
\author{
Michael Dietze
}
